Added calloc and realloc functions

+++ src/string/memset.c
@@ -3,7 +3,7 @@
 
 #include <string.h>
 
-void * memset (void *dest, int val, int len) {
+void * memset (void *dest, int val, size_t len) {
   unsigned char *ptr = dest;
   while (len-- > 0)
     *ptr++ = val;
